requirements.txt
python-resize-image
pillow
